ipa: ipu3: aiq: Provide managed AIQ Results structures
The AIQ results structures returned from the AIQ library has many
complex and deeply nested data structures which must be copied out.

This class handles the correct initialisation of a full set of AIQ
results, along with the allocations required, and the ability to copy
data into those structures correctly.

Signed-off-by: Umang Jain <umang.jain@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>